Add personal info script
This is the datavyu portion of the personal info censoring process.
This will output a csv file to the specified directory (specified in
the script) containing the timestamps for personal information onsets
and offsets, as well as their source (audio/video).

These comments should take the form:

%com: personal information [audio]: credit card number

or

%com: personal information [video]: butt

The output file should then be fed into the (soon to be) python program
videoscrub.py, which will do the actual video editing.