Have the script pull out the names of the columns itself
Instead of having the user write the name of the column that needs to
be processed, the script can now pull out the names of all the columns
in the current spreadsheet and run all the checks on each column. The
error messages will now display which column the error was in. This
might be useful during consensus, when there are two columns being
looked over. This commit will fix #1.