Output personalinfo.csv's into the same directory as working file
This commit changes the output location for the personalinfo csv files.
If there are personal info regions found in the current working file,
then those regions will be written to a csv file in the same directory
as the opf file that the script was run on. If there are no personal
info regions, then the name of the file will be appended to a list
found on Opus at /seedlings/Scripts_and_Apps/no_personal_info.txt. This
will serve as a running log of what files have been personalinfo
processed.